`test_covidcast.py` moved to integrations
